Creación de clase helper separando el método textToHtml y así poder cumplir con SRP
